backend/requirements.txt
# Web & Database
fastapi
uvicorn[standard]
asyncpg
httpx
pydantic
python-dotenv

# Authentication
python-jose[cryptography]
bcrypt
python-multipart
google-auth

# Agent Core
langchain
langgraph
langchain-community
langchain-core

# AI Models & Tools
langchain-google-genai
langchain-groq
langchain-tavily
tavily-python

# Development & CLI
langgraph-cli[inmem]